+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- [TD-2186] Return profile info in data structure view
 - [TD-2144] Support ngram-search in structure name
 - [TD-2159] Mapping for data field type